refactor: Feature Based Patternにディレクトリ構成を変えるために細かい箇所を分離
